A better wifi monitor.
Apple's wifi bars are a lie. I've opened up a radar, they're fine with
it being a lie. [0]

If you're signal is > -85 you're pretty much guaranteed to have five
bars, even if the noise is as high as the signal, which kills your
wifi and renders it useless.

They're probably doing this because you can't measure the noise without
joining the network. So a network that appeared 5 bars could drop to 0
bars when joined, this would confuse users. Solving it would be too Snow
Leopard-esque.

[0] http://openradar.appspot.com/radar?id=5032012842795008