Fixed a long standing annoyance with ssh!
Turns out: I needed to read the manual.

So ForwardAgent never seemed to work, even I set in a Host specific
block. It did work if I passed the `-A` flag to the ssh connection,
but otherwise it wouldn't.

Turns out: ssh uses the first value it finds as the value to use. So
because my `Host *` block came first, none of the ssh connections used
ForwardAgent unless I overrode it with the `-A` flag.